feat: enhance CI workflow and testing setup
- Added linting and unit testing steps to the GitHub Actions CI workflow.
- Updated AGENTS.md to clarify testing commands and conventions.
- Refactored ESLint configuration to use the latest Prettier config.
- Removed legacy Jest configuration file and replaced it with a modern ESM-based config.
- Updated package.json scripts for improved test command structure.
- Added integration tests for browser installation and uninstallation scenarios.
- Created integration tests for Qlik Sense Cloud thumbnail creation.
- Implemented error handling tests for browser uninstallation and thumbnail creation.
