+++ .github/agents/test-writer.md
@@ -54,6 +54,17 @@ describe("parseQVDHeader", () => {
 - Group related tests logically
 - Document non-obvious test setup
 
+## Test Type Convention (filename suffix)
+
+This project separates unit and integration tests by filename suffix:
+
+- `*.test.js` — **unit test**. Fast, no network, no external services, no credentials. Runs in `npm run test:unit` and `npm run test:watch`.
+- `*.integration.test.js` — **integration test**. Requires network access, real Qlik Sense servers (QSEoW or Cloud), certificates, browser downloads, or other external services. Runs in `npm run test:integration`.
+
+**Rule:** any test that needs network, credentials, real Qlik servers, browser downloads, or takes more than a few seconds MUST use the `.integration.test.js` suffix. This keeps `test:unit` fast and reliable for daily development.
+
+When in doubt, default to `.integration.test.js` — it's easy to move a test the other direction later, but a slow flaky test in the unit suite blocks every developer.
+
 ## When Writing Tests
 
 - Start with the critical path
