+++ .github/workflows/ci.yaml
@@ -7,7 +7,33 @@ on:
       - main
 
 jobs:
+  unit-tests:
+    permissions:
+      contents: read
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+        with:
+          fetch-depth: 2
+          persist-credentials: false
+
+      - name: Setup node
+        uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020 # v4
+        with:
+          node-version: 24
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Run lint
+        run: npm run lint
+
+      - name: Run unit tests
+        run: npm run test:unit
+
   test-on-macos-arm64:
+    needs:
+      - unit-tests
     permissions:
       contents: read
     concurrency:
@@ -72,7 +98,7 @@ jobs:
       - name: Run browser tests
         run: |
           ls -la
-          npm run test browser -- -i
+          npm run test:integration browser -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
@@ -96,7 +122,7 @@ jobs:
       - name: Run QS cloud tests
         run: |
           ls -la
-          npm run test cloud -- -i
+          npm run test:integration cloud -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
@@ -120,7 +146,7 @@ jobs:
       - name: Run QSEoW tests
         run: |
           ls -la
-          npm run test qseow -- -i
+          npm run test:integration qseow -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
@@ -215,7 +241,7 @@ jobs:
         shell: powershell
         run: |
           dir
-          npm run test browser -- -i
+          npm run test:integration browser -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
@@ -242,7 +268,7 @@ jobs:
         shell: powershell
         run: |
           dir
-          npm run test cloud -- -i
+          npm run test:integration cloud -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
@@ -269,7 +295,7 @@ jobs:
         shell: powershell
         run: |
           dir
-          npm run test qseow -- -i
+          npm run test:integration qseow -- -i
         working-directory: .
         env:
           BSI_LOG_LEVEL: 'info'
