+++ package.json
@@ -5,9 +5,11 @@
     "type": "module",
     "main": "src/butler-sheet-icons.js",
     "scripts": {
-        "test": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --runInBand --detectOpenHandles --forceExit --coverage",
-        "test:watch": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --runInBand --detectOpenHandles --forceExit --watch --coverage",
-        "test:unit": "npm test",
+        "jest": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --runInBand --detectOpenHandles --forceExit",
+        "test": "npm run test:unit && npm run test:integration",
+        "test:unit": "npm run jest -- --testPathIgnorePatterns='\\.integration\\.test\\.js$' --coverage",
+        "test:integration": "npm run jest -- --testPathPatterns='\\.integration\\.test\\.js$'",
+        "test:watch": "npm run test:unit -- --watch",
         "lint": "npx eslint src/ --max-warnings 0",
         "lint:fix": "npx eslint --fix src/",
         "format": "npx prettier --write \"src/**/*.js\"",
