+++ src/__tests__/jest.config.test.js
@@ -1,38 +1,28 @@
 // filepath: /Users/goran/code/butler-sheet-icons/src/__tests__/jest.config.test.js
 import { test, expect, describe } from '@jest/globals';
-import jestConfig from '../../jest.config.js';
+import jestConfig from '../../jest.config.mjs';
 
 describe('Jest Configuration', () => {
     test('should have correct basic configuration properties', () => {
         expect(jestConfig).toHaveProperty('clearMocks');
         expect(jestConfig).toHaveProperty('collectCoverage');
         expect(jestConfig).toHaveProperty('coverageDirectory');
+        expect(jestConfig).toHaveProperty('coverageProvider');
+        expect(jestConfig).toHaveProperty('testEnvironment', 'node');
         expect(jestConfig).toHaveProperty('roots');
     });
 
     test('should collect coverage correctly', () => {
         expect(jestConfig.collectCoverage).toBe(true);
-        expect(jestConfig.collectCoverageFrom).toContain('src/**/*.js');
         expect(jestConfig.coverageDirectory).toBe('coverage');
     });
 
-    test('should properly ignore paths for coverage', () => {
-        expect(jestConfig.coveragePathIgnorePatterns).toEqual(
-            expect.arrayContaining(['/node_modules/', '/build/', '/dist/'])
-        );
-    });
-
     test('should use correct root directory', () => {
-        expect(jestConfig.rootDir).toBe('.');
         expect(jestConfig.roots).toEqual(['<rootDir>/src']);
     });
 
-    test('should match test files correctly', () => {
-        expect(jestConfig.testMatch).toEqual(['**/__tests__/**/*.js', '**/src/**/*.test.js']);
-    });
-
-    test('should have appropriate transformation settings', () => {
-        expect(jestConfig.transform).toBeDefined();
-        expect(jestConfig.transformIgnorePatterns).toBeDefined();
+    test('should rely on Jest defaults for test discovery', () => {
+        // testMatch is unset; Jest defaults include __tests__/ folders and *.test.js files
+        expect(jestConfig.testMatch).toBeUndefined();
     });
 });
