+++ src/lib/browser/__tests__/browser_install_uninstall.integration.test.js
