+++ src/lib/browser/__tests__/browser_list_installed.integration.test.js
