+++ src/lib/browser/__tests__/browser_uninstall.integration.test.js
