+++ src/lib/cloud/__tests__/qscloudCreateThumbnails_success.integration.test.js
